+++ src/MediaServiceProvider.php
@@ -7,6 +7,7 @@ namespace NurbekJummayev\LaravelMediaApi;
 use Illuminate\Console\Scheduling\Schedule;
 use NurbekJummayev\LaravelMediaApi\Console\PurgeMediaCommand;
 use NurbekJummayev\LaravelMediaApi\Services\MediaService;
+use NurbekJummayev\LaravelMediaApi\Services\MediaUrlService;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -24,6 +25,7 @@ class MediaServiceProvider extends PackageServiceProvider
     public function packageRegistered(): void
     {
         $this->app->singleton(MediaService::class);
+        $this->app->singleton(MediaUrlService::class);
     }
 
     public function packageBooted(): void